SimulatedAnnealing Solver
added class and accompanying methods to candidatesolution to find an
optimal solution using simulated annealing.
it currently uses an integer value as a heat, and continually makes
changes to the solution until it makes that many bad changes (i.e.
changes that didn't lower the energy, and thus which are undone) in a
row.