added getSatisfaction method
added methods that rates overall how happy the students are with the
solution. for each student, it calculates how happy they would be with
it based off how far down their assigned project is on their list of
ordered preferences (i.e. 100% for top preference, 0% for bottom
preference).

the solution then finds the average satisfaction levels between all the
students